requirements.txt
Flask==3.0.3
openai==1.35.3
python-dotenv==1.0.1
gunicorn==22.0.0